pass target triple to zig builds
The zig build system by default selects a baseline CPU for the given target,
keeping the nix build deterministic without explicitely passing `-Dcpu=baseline`.

This also is a step in the right direction for cross compilation
because the target triple is made from stdenv's target platform
instead of defaulting to the host platform.

fixes nix builds for platforms that do not support `-Dcpu=baseline`