nix component: fix IFD build failure not recognized
There seems to be a bug in Nix that causes it to sometimes print an empty error message,
not allowing us to discern which builds actually failed by parsing it.
Recognition of a failure must not depend on the parsing of the error message
but instead solely on the exit code of the `nix build` process.