nix-sigstop: fix IPC bugs
Polling stopped on EOF from FIFO:
Due to `O_NDELAY`, `read()` returns 0 instead of blocking
when there is no data in the pipe, which `std.io.poll()`
handles by not polling that file descriptor anymore.
This lead to the pipe buffer filling up, blocking build hooks.

Message boundaries not properly handled:
Handle the case in which multiple complete messages are in the buffer.

No `done` message on failed build:
The build hook only sent the `done` message
after successfully building, not on failure,
causing the wrapper to wait for that message forever.